Add files via upload
You are building a shipping application for an online store. As part of the application, you need to create a Box class to handle box-related calculations. The Box class has a private attribute length, representing the length of the box.



To assist with volume calculations:

You decide to implement a constructor that sets the length attribute to 0. 
You also create a member function called get() that allows users to input the length value. 
Additionally, you include a friend function called printVolume(Box) that calculates and returns the box's volume.


Now, imagine a situation where a user wants to know the volume of a custom box they are considering for shipping their products. How would the user use the Box class to input the box's length and obtain the volume?

Input format :
The input consists of the length of the box N as an integer.

Output format :
The output prints the volume of the box.

Code constraints :
1 ≤ N ≤ 100

Sample test cases :
Input 1 :
15
Output 1 :
3375
Input 2 :
18
Output 2 :
5832