Add files via upload
You are developing a recipe management system that deals with fractional quantities of ingredients. To implement this functionality, you decide to create a Fraction class to handle fractional calculations. The Fraction class has private attributes numerator and a denominator to represent a fraction.



You need to implement a friend function addMixedFraction(Fraction& , Fraction& ) that takes two Fraction objects as parameters. One of the fractions may be a mixed fraction, where the numerator is greater than the denominator. The function should add the fractions together and return the result as a new Fraction object.



Write a program that allows users to input the numerators and denominators of two fractions. Create objects of the Fraction class using the input values. Then, use the addMixedFraction function to add the mixed fractions and display the result.



Example 1



Input:

4 7

8 7



Output:

12/7



Explanation:

The denominators of both fractions are the same (7), so the function directly adds the numerators and places the sum over the common denominator. Mathematically, 4/7 + 8/7 = (4 + 8) / 7 = 12 / 7.



Example 2



Input:

3 5

6 7



Output:

51/35



Explanation:

 As the denominators are different (5 and 7), it finds a common denominator (35) and calculates the sum of numerators: (3 * 7) + (6 * 5) = 21 + 30 = 51. So, 51/35 is printed.



Example 3



Input:

7 4

6 5



Output:

59/20



Explanation:

The denominators are different (4 and 5), so a common denominator is found by multiplying them (20). The numerators are calculated by cross-multiplying and adding: (7 * 5) + (6 * 4) = 35 + 24 = 59. So, 59/20 is printed.



Note: This is a sample question asked in Capgemini recruitment.

Input format :
The first line consists of two space-separated integers, representing the numerator and denominator of the first fraction.

The second line consists of two space-separated integers, representing the numerator and denominator of the second fraction.

Output format :
The output displays the result of adding the two fractions in the following format: numerator/denominator.



Refer to the sample output for the formatting specifications.

Code constraints :
1 ≤ numerator ≤ 2000

1 ≤ denominator ≤ 100

Denominators will not be zero

Sample test cases :
Input 1 :
4 7
8 7
Output 1 :
12/7
Input 2 :
3 5
6 7
Output 2 :
51/35
Input 3 :
7 4
6 5
Output 3 :
59/20