Add files via upload
John is a math enthusiast who loves to create programs that perform various mathematical operations. He recently decided to write a program to calculate the square, modulus, and division of numbers. He created three functions with the same name, "calc," each serving a different purpose.



Firstly, John implemented the function for calculating the square whenever he passed an integer argument to this function and returned the square of that number as an integer value.



Next, John developed the function for calculating the modulus. This function takes in two integer arguments and calculates the modulus of the first number divided by the second number. It then returns the modulus as an integer value.



Lastly, John implemented the function for performing division. This function takes in two floating-point arguments and performs the division operation on them. It returns the result as a float value.



John's program allows users to perform these mathematical operations by simply calling the "calc" function with appropriate arguments, making it easier for them to obtain the desired results without writing extensive code.



Note: This is a sample question asked in an Infosys interview.

Input format :
The input consists of a single line input that has 2 integer values and 2 floating-point values, separated by a space.



Note:

First input for Square function
First and second inputs for modulus function
Last two inputs for division function
Output format :
The output displays the result of the operations in separate lines.



Refer to the sample output for formatting specifications.

Sample test cases :
Input 1 :
2 5 6.3 5.2
Output 1 :
Square of 2: 4
Modulus of 2, 5: 2
Quotient of 6.3, 5.2: 1.21